+++ ChangeLog.md
@@ -1,5 +1,7 @@
 # Changes in 1.1.7
 
+* Add a [factory](incl/../include/clap/factory) folder for better organization.
+
 ## Draft extensions
 
 * [preset-discovery.h](include/clap/preset-discovery.h): new extension which allows the host to index the plugin's preset which are stored on disk.